cleanup and refactor - wip
